Framework for Behaviour Trees Complete
Completed writing the scripts that form the framework of a Behaviour Tree.
Previously the BTNode.cs contains all the script for Behaviour Tree Node creation.
Tree.cs contains a reference to the root node and builds a tree according to how it is defined in the SetupTree method.
Selector.cs and Sequence.cs are composite nodes crucial in any behaviour tree and function like the OR and AND logic gates respectively.